Add front end functionality to edit orgs in admin > #351
